requirements.txt
setuptools~=75.8.0
click~=8.1.8
python-dotenv~=1.0.1
protobuf~=5.29.3